+++ lib/load.php
@@ -146,6 +146,9 @@ if ( gutenberg_is_experiment_enabled( 'gutenberg-dataform-inspector' ) ) {
 if ( gutenberg_is_experiment_enabled( 'gutenberg-media-editor' ) ) {
 	require __DIR__ . '/experimental/media-editor/load.php';
 }
+if ( gutenberg_is_experiment_enabled( 'gutenberg-media-folders' ) ) {
+	require __DIR__ . '/experimental/media-folders.php';
+}
 
 if ( gutenberg_is_experiment_enabled( 'gutenberg-workflow-palette' ) ) {
 	require __DIR__ . '/experimental/workflow-palette.php';