+++ packages/block-editor/src/store/private-actions.js
@@ -15,6 +15,7 @@ const castArray = ( maybeArray ) =>
  */
 const privateSettings = [
 	'inserterMediaCategories',
+	'inserterMediaFolders',
 	'blockInspectorAnimation',
 	'mediaSideload',
 ];